online-serving-assign/pyproject.toml
[tool.poetry]
name = "web-single"
version = "0.1.0"
description = ""
authors = ["heumsi <heumsi@naver.com>"]
readme = "README.md"
packages = [{include = "web_single"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.105.0"}
scikit-learn = "^1.3.2"
sqlmodel = "^0.0.14"
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.1"
pillow = "^10.2.0"
torch = {version = "2.0.1+cu117", source = "pytorch"}
torchvision = {version = "0.15.2+cu117", source = "pytorch"}


[tool.poetry.group.dev.dependencies]
pandas = "^2.1.4"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu117"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
